// src/features/persona/models/constants.ts
/** 선택 유형 */
enum ChoiceType {
  SingleChoice = "SingleChoice",
  MultipleChoice = "MultipleChoice",
}
/** 질문 카테고리 */
enum QuestionCategory {
  Normal = "Normal",
  Welfare = "Welfare",
  Transportation = "Transportation",
  Economical = "Economical",
}

export { ChoiceType, QuestionCategory };